Add correct while loop structure to program
The previous iteration ran one argument-less command as an argument
from the command line, which is not what the assignment required. This
basic functionality has now been reproduced through a while loop as
the interactive mode.

Specific errors have been removed and replaced with a single generic
error wrapped in a function.