Add batch mode
Instead of creating a new function for the batch mode, I adapted the
interactive mode function to receive an input stream as an
argument. This required a few adjustments, such as making sure that
the terminal prompt is only printed when the input stream is using
stdin.

A few changes to exit and return statements were made and may be
further adjusted because it's not entirely clear to me yet when which
would be more appropriate.

command_w_args uses a fix buffer size for each string now. I'm not
entirely use when that became necessary, but at least it reminds me
that I may need to be thinking about string size and memory more
carefully in this program.