Refactor redirection logic
In order to better handle each piece of a command--especially in
anticipation of implementing parallel commands--commands are now put
into a struct. There structs are created before checking the built-in
commands and elimiate the clumsy mess of endless else/if blocks that
followed the built-ins. It's still a bit clumsy, but more siginicantly
more legible to me.

The issue mentioned in the last commit was simply because of a mistake
in the trimWhiteSpace function.